// src/config/enum.ts
export enum USER_TYPE {
  BOTH = '3',
  GUEST = '1',
  REGISTERED = '2'
}

export enum CHANNEL_TYPE {
  NOTICIEROS = 'noticieros',
  FORO_TV = 'foro-tv'
}

export enum PLATFORM_TYPE {
  MOBILE = 'MOBILE',
  BOTH = 'BOTH'
}

export enum COLLECTION_TYPE {
  POSTS = 'POSTS',
  VIDEOS = 'VIDEOS',
  LIVE_BLOGS = 'LIVE_BLOGS',
  PROGRAMS = 'PROGRAMS',
  TALENTS = 'TALENTS',
  PODCASTS = 'PODCASTS',
  PRESS_ROOM = 'PRESS_ROOM',
  INTERACTIVOS = 'PAGES',
  AUTHORS = 'AUTHORS'
}

export enum SORTING_TYPE {
  MOST_RECENT = 'RECENT',
  RELEVANCE = 'RELEVANCE'
}

export enum SocialProvider {
  APPLE = '1',
  GOOGLE = '2',
  FACEBOOK = '3'
}